Q&A ComfirmDelete

+++ src/main/webapp/common/main/main.js
@@ -31,15 +31,17 @@ function updateCart(obj, id){
 
 
 function deleteCart(id){
-	fetch(`http://localhost:8080/shop-style/home-page/api/cart/${id}`,{
-		method: "DELETE"
-	}).then(function(res){
-		return res.json() 
-	}).then(function(data){
-		var d = document.getElementById("cartCouter");
-		if(d != null){
-			d.innerHTML = data
-			location.reload()
-		}
-	})
+	if(confirm("Bạn có muốn xóa không ?") == true){
+		fetch(`http://localhost:8080/shop-style/home-page/api/cart/${id}`,{
+			method: "DELETE"
+		}).then(function(res){
+			return res.json() 
+		}).then(function(data){
+			var d = document.getElementById("cartCouter");
+			if(d != null){
+				d.innerHTML = data
+				location.reload()
+			}
+		})
+	}
 }
